Move BottomLine to its own file
